feat : signIn

+++ src/entity/post.entity.ts
@@ -3,7 +3,7 @@ import {Column, Entity, PrimaryGeneratedColumn } from 'typeorm'
 @Entity()
 export class Post {
     @PrimaryGeneratedColumn({
-        type: "bigint"
+        type: "int"
     })
     postId!: number
